adjusted some models

+++ workoutAPI/Models/Auth/ApiKey/ApiKeyDTO.cs
@@ -23,7 +23,7 @@ public class ApiKeyDTO :BaseModel
     public bool IsActive { get; set; }
     
     [Column("req_today")]
-    public int ReqToday { get; set; }
+    public decimal ReqToday { get; set; }
     
     [Column("last_reset_date")]
     public DateTime LastResetDate { get; set; }
